+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+- Customisation of subtype names with `@SubTypeName`.
 
 ## 0.1.6 2018-05-13
 ### Added